+++ sampledb/logic/schemas/generate_placeholder.py
@@ -37,6 +37,8 @@ def generate_placeholder(schema: dict, path: typing.Optional[typing.List[str]] =
         return _generate_quantity_placeholder(schema, path)
     elif schema['type'] == 'sample':
         return _generate_sample_placeholder(schema, path)
+    elif schema['type'] == 'measurement':
+        return _generate_measurement_placeholder(schema, path)
     elif schema['type'] == 'tags':
         return _generate_tags_placeholder(schema, path)
     elif schema['type'] == 'hazards':
@@ -198,5 +200,16 @@ def _generate_sample_placeholder(schema: dict, path: typing.List[str]) -> None:
     :param schema: the sampledb object schema
     :param path: the path to this subschema
     :return: None, as there can be no default sample
+    """
+    return None
+
+
+def _generate_measurement_placeholder(schema: dict, path: typing.List[str]) -> None:
+    """
+    Generates a placeholder measurement object based on an object schema.
+
+    :param schema: the sampledb object schema
+    :param path: the path to this subschema
+    :return: None, as there can be no default measurement
     """
     return None