+++ sampledb/logic/schemas/validate_schema.py
@@ -52,6 +52,8 @@ def validate_schema(schema: dict, path: typing.Optional[typing.List[str]] = None
         return _validate_quantity_schema(schema, path)
     elif schema['type'] == 'sample':
         return _validate_sample_schema(schema, path)
+    elif schema['type'] == 'measurement':
+        return _validate_measurement_schema(schema, path)
     elif schema['type'] == 'tags':
         return _validate_tags_schema(schema, path)
     elif schema['type'] == 'hazards':
@@ -368,6 +370,27 @@ def _validate_quantity_schema(schema: dict, path: typing.List[str]) -> None:
 def _validate_sample_schema(schema: dict, path: typing.List[str]) -> None:
     """
     Validates the given sample object schema and raises a ValidationError if it is invalid.
+
+    :param schema: the sampledb object schema
+    :param path: the path to this subschema
+    :raise ValidationError: if the schema is invalid.
+    """
+    valid_keys = {'type', 'title', 'note'}
+    required_keys = {'type', 'title'}
+    schema_keys = set(schema.keys())
+    invalid_keys = schema_keys - valid_keys
+    if invalid_keys:
+        raise ValidationError('unexpected keys in schema: {}'.format(invalid_keys), path)
+    missing_keys = required_keys - schema_keys
+    if missing_keys:
+        raise ValidationError('missing keys in schema: {}'.format(missing_keys), path)
+    if 'note' in schema and not isinstance(schema['note'], str):
+        raise ValidationError('note must be str', path)
+
+
+def _validate_measurement_schema(schema: dict, path: typing.List[str]) -> None:
+    """
+    Validates the given measurement object schema and raises a ValidationError if it is invalid.
 
     :param schema: the sampledb object schema
     :param path: the path to this subschema