+++ tests/logic/schemas/test_validate.py
@@ -1019,6 +1019,193 @@ def test_validate_sample_invalid_object_id():
         validate(instance, schema)
 
 
+def test_validate_measurement():
+    from sampledb.models.users import User, UserType
+    from sampledb.models.actions import Action, ActionType
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    user = User("User", "example@fz-juelich.de", UserType.OTHER)
+    action = Action(ActionType.MEASUREMENT, "Example Action", schema={
+      "title": "Measurement Information",
+      "type": "object",
+      "properties": {
+        "name": {
+          "title": "Measurement Name",
+          "type": "text"
+        }
+      },
+      'required': ['name']
+    })
+
+    sampledb.db.session.add(user)
+    sampledb.db.session.add(action)
+    sampledb.db.session.commit()
+
+    object = create_object(data={'name': {'_type': 'text', 'text': 'example'}}, user_id=user.id, action_id=action.id)
+    instance = {
+        '_type': 'measurement',
+        'object_id': object.id
+    }
+    validate(instance, schema)
+
+
+def test_validate_measurement_invalid_type():
+    from sampledb.models.users import User, UserType
+    from sampledb.models.actions import Action, ActionType
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    user = User("User", "example@fz-juelich.de", UserType.OTHER)
+    action = Action(ActionType.MEASUREMENT, "Example Action", schema={
+      "title": "Measurement Information",
+      "type": "object",
+      "properties": {
+        "name": {
+          "title": "Measurement Name",
+          "type": "text"
+        }
+      },
+      'required': ['name']
+    })
+
+    sampledb.db.session.add(user)
+    sampledb.db.session.add(action)
+    sampledb.db.session.commit()
+
+    object_id = create_object(data={'name': {'_type': 'text', 'text': 'example'}}, user_id=user.id, action_id=action.id)
+    instance = object_id
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
+def test_validate_measurement_unexpected_keys():
+    from sampledb.models.users import User, UserType
+    from sampledb.models.actions import Action, ActionType
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    user = User("User", "example@fz-juelich.de", UserType.OTHER)
+    action = Action(ActionType.MEASUREMENT, "Example Action", schema={
+      "title": "Measurement Information",
+      "type": "object",
+      "properties": {
+        "name": {
+          "title": "Measurement Name",
+          "type": "text"
+        }
+      },
+      'required': ['name']
+    })
+
+    sampledb.db.session.add(user)
+    sampledb.db.session.add(action)
+    sampledb.db.session.commit()
+
+    object_id = create_object(data={'name': {'_type': 'text', 'text': 'example'}}, user_id=user.id, action_id=action.id)
+    instance = {
+        '_type': 'measurement',
+        'object_id': object_id,
+        'action_id': action.id
+    }
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
+def test_validate_measurement_missing_keys():
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    instance = {
+        '_type': 'measurement'
+    }
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
+def test_validate_measurement_wrong_type():
+    from sampledb.models.users import User, UserType
+    from sampledb.models.actions import Action, ActionType
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    user = User("User", "example@fz-juelich.de", UserType.OTHER)
+    action = Action(ActionType.MEASUREMENT, "Example Action", schema={
+      "title": "Measurement Information",
+      "type": "object",
+      "properties": {
+        "name": {
+          "title": "Measurement Name",
+          "type": "text"
+        }
+      },
+      "required": ["name"]
+    })
+
+    sampledb.db.session.add(user)
+    sampledb.db.session.add(action)
+    sampledb.db.session.commit()
+
+    object_id = create_object(data={'name': {'_type': 'text', 'text': 'example'}}, user_id=user.id, action_id=action.id)
+    instance = {
+        '_type': 'object_reference',
+        'object_id': object_id
+    }
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
+def test_validate_measurement_wrong_object_id_type():
+    from sampledb.models.users import User, UserType
+    from sampledb.models.actions import Action, ActionType
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    user = User("User", "example@fz-juelich.de", UserType.OTHER)
+    action = Action(ActionType.MEASUREMENT, "Example Action", schema={
+      "title": "Measurement Information",
+      "type": "object",
+      "properties": {
+        "name": {
+          "title": "Measurement Name",
+          "type": "text"
+        }
+      },
+      "required": ["name"]
+    })
+
+    sampledb.db.session.add(user)
+    sampledb.db.session.add(action)
+    sampledb.db.session.commit()
+
+    object = create_object(data={'name': {'_type': 'text', 'text': 'example'}}, user_id=user.id, action_id=action.id)
+    instance = {
+        '_type': 'measurement',
+        'object_id': object
+    }
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
+def test_validate_measurement_invalid_object_id():
+    schema = {
+        'title': 'Example',
+        'type': 'measurement'
+    }
+    instance = {
+        '_type': 'measurement',
+        'object_id': 42
+    }
+    with pytest.raises(ValidationError):
+        validate(instance, schema)
+
+
 def test_validate_tags():
     schema = {
         'title': 'Example',